fix: this hack should delete all instances of server on windows

+++ src/electron/main.ts
@@ -204,6 +204,17 @@ function handleCloseEvents(
 
 function quitServerProcess(serverProcess: ChildProcess | null) {
     if (serverProcess) {
-        serverProcess.kill();
+        // Kill entire process tree on Windows.
+        if (process.platform === "win32" && serverProcess.pid) {
+            spawn(
+                "taskkill",
+                ["/pid", serverProcess.pid.toString(), "/t", "/f"],
+                {
+                    stdio: "ignore",
+                }
+            );
+        } else {
+            serverProcess.kill();
+        }
     }
 }
